返回的内容添加了category。distance会在结束运动时更新。

+++ src/main/java/edu/tongji/backend/entity/Exercise.java
@@ -24,5 +24,5 @@ public class Exercise {
     private String category;
 
     private int exerciseId;
-    private Integer distance;
+    private double distance;
 }
